Log sorting
Logs from the Doppler log streaming endpoint are not guaranteed to be in order
temporally.  Previously, the log streaming operation (Applications#logs)
punted on this issue, requiring the user to deal with it.  Thanks to Kris De
Volder doing the hard work of designing a solid reactive algorithm, we can now
include this functionality in the client itself.